refactor: Header 영역 컴포넌트 그룹화 및 폴더 구조 변경
