+++ src/components/Header/Logo.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import shoppingCartIcon from "../asset/shopping-cart-icon.svg";
+import shoppingCartIcon from "../../asset/shopping-cart-icon.svg";
 
 const LogoContainer = styled.div`
   display: flex;
