+++ src/stories/Header/Header.stories.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import Header from "../../components/Header";
+
+export default {
+  title: "Component/Header",
+  component: Header,
+};
+
+const Template = (args) => <Header {...args} />;
+
+export const Default = Template.bind({});
