+++ src/stories/Header/Logo.stories.jsx
@@ -1,17 +1,11 @@
 import React from "react";
-import Logo from "../components/Logo";
+import Logo from "../../components/Header/Logo";
 
 export default {
   title: "Component/Logo",
   component: Logo,
-  argTypes: {
-    children: { controls: "text" },
-  },
 };
 
 const Template = (args) => <Logo {...args} />;
 
 export const Default = Template.bind({});
-Default.args = {
-  children: "안녕",
-};
